fix(FOROME-693): gene region
