+++ package.json
@@ -10,7 +10,6 @@
     "chart.js": "3.5.1",
     "date-fns": "2.21.1",
     "faker": "5.5.3",
-    "formik": "2.2.8",
     "igv": "2.10.5",
     "lodash": "4.17.21",
     "md5": "2.3.0",
