+++ src/pages/filter/ui/panels/function-panel/function-panel.tsx
@@ -4,7 +4,7 @@ import filterStore from '@store/filter'
 import { CompundHet } from './components/compound-het/compound-het'
 import { CompoundRequest } from './components/compound-request/compound-request'
 import { CustomInheritanceMode } from './components/custom-inheritance-mode/custom-inheritance-mode'
-import { GeneRegion } from './components/gene-region'
+import { GeneRegion } from './components/gene-region/gene-region'
 import { InheritanceMode } from './components/inheritance-mode/inheritance-mode'
 
 const functionsMap: Record<string, any> = {
